+++ caom2-pkg-server/build.gradle
@@ -13,7 +13,7 @@ sourceCompatibility = 1.8
 
 group = 'org.opencadc'
 
-version = '1.7.1'
+version = '1.8.0'
 
 description = 'OpenCADC CAOM2 package server library'
 def git_url = 'https://github.com/opencadc/caom2service'
@@ -30,6 +30,7 @@ dependencies {
     compile 'org.opencadc:cadc-uws-server:[1.2,1.3)'
     compile 'org.opencadc:caom2:[2.4,2.5)'
     compile 'org.opencadc:caom2-tap:[1.7,)'
+    compile 'org.opencadc:cadc-pkg-server:[1.0.0,)'
 
     testCompile 'junit:junit:[4.0,5.0)'
 }
