+++ caom2-pkg-server/src/test/java/ca/nrc/cadc/caom2/pkg/Caom2PackageRunnerTest.java
@@ -3,7 +3,7 @@
 *******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 **************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 *
-*  (c) 2018.                            (c) 2018.
+*  (c) 2021.                            (c) 2021.
 *  Government of Canada                 Gouvernement du Canada
 *  National Research Council            Conseil national de recherches
 *  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -81,21 +81,22 @@ import org.junit.Test;
  *
  * @author pdowler
  */
-public class PackageRunnerTest {
-    private static final Logger log = Logger.getLogger(PackageRunnerTest.class);
+public class Caom2PackageRunnerTest {
+    private static final Logger log = Logger.getLogger(Caom2PackageRunnerTest.class);
 
     static {
         Log4jInit.setLevel("ca.nrc.cadc.caom2", Level.INFO);
+        Log4jInit.setLevel("org.opencadc.pkg.server", Level.INFO);
     }
     
-    public PackageRunnerTest() { 
+    public Caom2PackageRunnerTest() {
     }
     
     @Test
     public void testStandardPublisherID() {
         try {
             PublisherID pid = new PublisherID(URI.create("ivo://cadc.nrc.ca/COLLECTION?observationID/productID"));
-            PlaneURI puri = PackageRunner.toPlaneURI(pid);
+            PlaneURI puri = Caom2PackageRunner.toPlaneURI(pid);
             log.info(pid.getURI() + " -> " + puri);
             Assert.assertNotNull(puri);
         } catch (Exception unexpected) {
@@ -108,7 +109,7 @@ public class PackageRunnerTest {
     public void testPrefixedPublisherID() {
         try {
             PublisherID pid = new PublisherID(URI.create("ivo://cadc.nrc.ca/prefix/COLLECTION?observationID/productID"));
-            PlaneURI puri = PackageRunner.toPlaneURI(pid);
+            PlaneURI puri = Caom2PackageRunner.toPlaneURI(pid);
             log.info(pid.getURI() + " -> " + puri);
             Assert.assertNotNull(puri);
         } catch (Exception unexpected) {
